ch ch changes

+++ Record Store/WebApplication1/Controllers/OrdersController.cs
@@ -1,8 +1,12 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
+using Record_Store.Auth;
 using Record_Store.Data;
 using Record_Store.Data.DTOS.Orders;
 using Record_Store.Data.Repositories;
 using Record_Store.Entity;
+using Microsoft.IdentityModel.JsonWebTokens;
+using System.Security.Claims;
 using System.Text.Json;
 
 namespace Record_Store.Controllers
@@ -12,11 +16,13 @@ namespace Record_Store.Controllers
     public class OrdersController : ControllerBase
     {
         private readonly IOrdersRepository _ordersRepository;
-        public OrdersController(IOrdersRepository ordersRepository)
+        private readonly IAuthorizationService _authorizationService;
+        public OrdersController(IOrdersRepository ordersRepository, IAuthorizationService authorizationService)
         {
             _ordersRepository= ordersRepository;
+            _authorizationService=authorizationService;
         }
-        //[HttpGet]
+        [HttpGet]
         public async Task<IEnumerable<OrderDTO>> GetMany()
         {
             var orders = await _ordersRepository.GetOrdersManyAsync();
@@ -24,27 +30,27 @@ namespace Record_Store.Controllers
             return orders.Select(o => new OrderDTO(o.ID, o.Name, o.Price, o.CreatedDate));
         }
 
-        [HttpGet(Name = "GetOrders")]
-        public async Task<IEnumerable<OrderDTO>> GetManyPaging([FromQuery] SearchParameters searchParameters)
-        {
-            var orders = await _ordersRepository.GetOrdersManyPagedAsync(searchParameters);
+        //[HttpGet(Name = "GetOrders")]
+        //public async Task<IEnumerable<OrderDTO>> GetManyPaging([FromQuery] SearchParameters searchParameters)
+        //{
+        //    var orders = await _ordersRepository.GetOrdersManyPagedAsync(searchParameters);
 
-            var previousPage = orders.hasPrevious ? CreateOrdersResourceUri(searchParameters, ResourceUriType.PreviousPage) : null;
-            var nextPage = orders.hasNext ? CreateOrdersResourceUri(searchParameters, ResourceUriType.NextPage) : null;
+        //    var previousPage = orders.hasPrevious ? CreateOrdersResourceUri(searchParameters, ResourceUriType.PreviousPage) : null;
+        //    var nextPage = orders.hasNext ? CreateOrdersResourceUri(searchParameters, ResourceUriType.NextPage) : null;
 
-            var paginationMetaData = new
-            {
-                totalCount = orders.TotalCount,
-                pageSize = orders.PageSize,
-                currentPage = orders.CurrentPage,
-                totalPages = orders.TotalPages,
-                previousPage,
-                nextPage
-            };
+        //    var paginationMetaData = new
+        //    {
+        //        totalCount = orders.TotalCount,
+        //        pageSize = orders.PageSize,
+        //        currentPage = orders.CurrentPage,
+        //        totalPages = orders.TotalPages,
+        //        previousPage,
+        //        nextPage
+        //    };
 
-            Response.Headers.Add("Pagination", JsonSerializer.Serialize(paginationMetaData));
-            return orders.Select(o => new OrderDTO(o.ID, o.Name, o.Price, o.CreatedDate));
-        }
+        //    Response.Headers.Add("Pagination", JsonSerializer.Serialize(paginationMetaData));
+        //    return orders.Select(o => new OrderDTO(o.ID, o.Name, o.Price, o.CreatedDate));
+        //}
 
         [HttpGet("{orderID}", Name = "GetOrder")]
         public async Task<ActionResult<OrderDTO>> Get(uint orderID)
@@ -77,10 +83,18 @@ namespace Record_Store.Controllers
         //}
 
         [HttpPost]
+        [Authorize(Roles = StoreRoles.StoreUser)]
         public async Task<ActionResult<OrderDTO>> Create(CreateOrderDTO createOrderDTO)
         {
-            var order = new Order { Name = createOrderDTO.Name, Price=createOrderDTO.Price, CreatedDate=DateTime.UtcNow, IsActive = true};
+            var order = new Order { Name = createOrderDTO.Name, Price=createOrderDTO.Price, CreatedDate=DateTime.UtcNow, IsActive = true, UserId = User.FindFirstValue(JwtRegisteredClaimNames.Sub) };
             await _ordersRepository.CreateOrder(order);
+
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, order, PolicyNames.ResourceOwner);
+            if (!authorizationResult.Succeeded)
+            {
+                // 404
+                return Forbid();
+            }
 
             //return CreatedAtAction("GetOrder", new { orderID = order.ID }, new OrderDTO(order.Name, order.Price,order.CreatedDate));
             //return CreatedAtAction(nameof(Get), new OrderDTO(order.Name, order.Price, order.CreatedDate));
@@ -88,6 +102,7 @@ namespace Record_Store.Controllers
         }
 
         [HttpPut]
+        [Authorize(Roles = StoreRoles.StoreUser)]
         [Route("{orderID}")]
         public async Task<ActionResult<OrderDTO>> Update(uint orderID, UpdateOrderDTO updateOrderDTO)
         {
@@ -97,7 +112,24 @@ namespace Record_Store.Controllers
             {
                 return NotFound();
             }
+
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, order, PolicyNames.ResourceOwner);
+            if (!authorizationResult.Succeeded)
+            {
+                // 404
+                return Forbid();
+            }
+
             order.Name = updateOrderDTO.Name;
+
+            if(order.DeliveryDate >= order.CreatedDate)
+            {
+                order.IsActive=false;
+            }
+            else
+            {
+                order.IsActive=true;
+            }
             await _ordersRepository.UpdateOrder(order);
             return Ok(new OrderDTO(order.ID, order.Name, order.Price, order.CreatedDate));
         }
@@ -112,6 +144,14 @@ namespace Record_Store.Controllers
             {
                 return NotFound();
             }
+
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, order, PolicyNames.ResourceOwner);
+            if (!authorizationResult.Succeeded)
+            {
+                // 404
+                return Forbid();
+            }
+
             await _ordersRepository.RemoveOrder(order);
 
             return NoContent();
